// frontend/src/api/generated/model/aiotBackendApiV1RuleNodeDefinition.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Nunu Example API
 * This is a sample server celler server.
 * OpenAPI spec version: 1.0.0
 */

export interface AiotBackendApiV1RuleNodeDefinition {
  category?: string
  configSchema?: string
  defaultConfig?: string
  description?: string
  executorKey?: string
  icon?: string
  inputPorts?: string
  isSystem?: boolean
  key?: string
  name?: string
  outputPorts?: string
  uuid?: string
}